groq_server/requirements.txt
fastmcp>=0.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
numpy>=1.21.0
asyncio>=3.4.3
